Get User Get School Get Accounts All are saved using NSUserDefaults

+++ DanceWorks Mobile/Account.h
@@ -15,13 +15,13 @@
 }
 
 
-//Decalres public property as well as wether to store a copy or a pointer of this variable in the model
+//Declares public property as well as wether to store a copy or a pointer of this variable in the model
 /* Properties
  *Copy - Required when object is mutuable. Creates a copy of it to be changed;
  *Assign - used for primitive types. Returns a reference to the object
  */
-@property (assign, nonatomic) int SchID;
-@property (assign, nonatomic) int AcctID;
+@property (assign, nonatomic) NSInteger SchID;
+@property (assign, nonatomic) NSInteger AcctID;
 @property (assign, nonatomic) NSString *AcctNo;
 @property (assign, nonatomic) NSString *FName;
 @property (assign, nonatomic) NSString *LName;
@@ -33,17 +33,17 @@
 @property (assign, nonatomic) NSString *EMail;
 @property () NSDate *DateReg;
 @property () NSDate *InactiveDate;
-@property (assign, nonatomic) int Status;
-@property (assign, nonatomic) int BillingFreq;
+@property (assign, nonatomic) NSInteger Status;
+@property (assign, nonatomic) NSInteger BillingFreq;
 @property (assign, nonatomic) float Balance;
 @property () NSDate *LastPmtDate;
 @property (assign, nonatomic) float LastPmtAmt;
 @property (assign, nonatomic) float NoStudents;
-@property (assign, nonatomic) int TuitionSel;
+@property (assign, nonatomic) NSInteger TuitionSel;
 @property (assign, nonatomic) NSString *Alert;
 @property (assign, nonatomic) NSString *Source;
 @property (assign, nonatomic) float MTuition;
-@property (assign, nonatomic) int ClassTime;
+@property (assign, nonatomic) NSInteger ClassTime;
 @property (assign, nonatomic) NSString *P1Name;
 @property (assign, nonatomic) NSString *P1EMail;
 @property (assign, nonatomic) NSString *P1Phone;
@@ -64,7 +64,7 @@
  @property (assign, nonatomic) NSString *P4Phone;
  @property (assign, nonatomic) NSString *P4Cell;
  @property (assign, nonatomic) NSString *P4Work;*/
-@property (assign, nonatomic) int CCType;
+@property (assign, nonatomic) NSInteger CCType;
 @property (assign, nonatomic) NSString *CCTrail;
 @property (assign, nonatomic) NSString *CCExpire;
 @property (assign, nonatomic) NSString *CCFName;
@@ -73,9 +73,9 @@
 @property (assign, nonatomic) NSString *CCCity;
 @property (assign, nonatomic) NSString *CCState;
 @property (assign, nonatomic) NSString *CCZip;
-@property (assign, nonatomic) int CCConsentID;
+@property (assign, nonatomic) NSInteger CCConsentID;
 @property (assign, nonatomic) BOOL CCMonthly;
-@property (assign, nonatomic) int CCDate;
+@property (assign, nonatomic) NSInteger CCDate;
 @property (assign, nonatomic) NSString *Notes;
 @property (assign, nonatomic) NSString *AccountFee;
 @property (assign, nonatomic) float AccountFeeAmount;
@@ -87,8 +87,8 @@
 @property (assign, nonatomic) NSString *ResetString;
 @property (assign, nonatomic) BOOL Activated;
 @property (assign, nonatomic) BOOL UsesNewPassword;
-@property (assign, nonatomic) int AlertID;
-@property (assign, nonatomic) int SourceID;
+@property (assign, nonatomic) NSInteger AlertID;
+@property (assign, nonatomic) NSInteger SourceID;
 @property (assign, nonatomic) BOOL Subscribed;
 @property (assign, nonatomic) NSString *CCToken;
